+++ lib/sources/consts/semantic_labels.dart
@@ -1,0 +1,25 @@
+class SemanticLabels {
+  static const rowLabel1 = "row_1";
+  static const rowLabel2 = "row_2";
+  static const rowLabel3 = "row_3";
+  static const rowLabel4 = "row_4";
+  static const rowLabel5 = "row_5";
+  static const rowLabel6 = "row_6";
+  static const rowLabel7 = "row_7";
+  static const rowLabel8 = "row_8";
+  static const rowLabel9 = "row_9";
+  static const rowLabel10 = "row_10";
+
+  static List<String> rowLabels = [
+    rowLabel1,
+    rowLabel2,
+    rowLabel3,
+    rowLabel4,
+    rowLabel5,
+    rowLabel6,
+    rowLabel7,
+    rowLabel8,
+    rowLabel9,
+    rowLabel10,
+  ];
+}